requirements.txt
pytest
pytest-xdist
selenium
os
time
logging
allure-pytest
shutil
webdriver-manager
requests
appium-python-client
